LayoutTests/media/video-src-set-expected.txt

Test that setting src attribute triggers load
EVENT(loadstart)
EXPECTED (relativeURL(video.currentSrc) == 'content/test.mp4') OK
END OF TEST
